+++ clear.sh
@@ -1,2 +1,3 @@
-rm -rf .venv .pytest_cache/ .ropeproject/ */__pycache__ build/ dist/ test/roms/ test/roms_dest/mastersystem/
+rm -rf .venv .pytest_cache/ .ropeproject/ */__pycache__ build/ dist/
+# rm -rf test/roms_dest/* test/roms/
 rm proccessing_gifs/Thumbs.db proccessing_gifs/Thumbs.db:encryptable *.spec