+++ sbfrm_gui.py
@@ -5,6 +5,9 @@ import threading
 from os.path import basename, join
 import PySimpleGUI as sg
 
+import json
+
+from src.fileutils import configs
 from src.collection import Collection
 
 
@@ -13,74 +16,65 @@ def make_window():
     last_up_col = sg.user_settings_get_entry('-last up_col-', '')
     last_up_sys = sg.user_settings_get_entry('-last up_sys-', '')
     if not last_up_col and not last_up_sys:
-        last_up_col = True
-        last_up_sys = False
-    print(last_up_col, last_up_sys)
-    operation = [[sg.Radio('Update Collection', "-RADIO-",
-                           default=last_up_col, key="-UPDATE_COLLECTION-"),
-                  sg.Radio('Update System', "-RADIO-",
-                           default=last_up_sys, key="-UPDATE_SYSTEM-")]]
+        last_up_col = False
+        last_up_sys = True
+    
+    choose_operation = [sg.Text("Choose Operation.", font=("Helvetica", 12))]
+    
+    operation = [[sg.Radio('Update System', "-RADIO-",
+                           default=last_up_sys, key="-UPDATE_SYSTEM-"),
+                 sg.Radio('Update Collection', "-RADIO-",
+                           default=last_up_col, key="-UPDATE_COLLECTION-")]]
+    
+    set_src_dest = [sg.Text("Set Source and Destination Paths.", font=("Helvetica", 12))]
 
-    last_src = sg.user_settings_get_entry('-last source-', '')
     select_src = [sg.Text("     Source", size=(13, 1)),
-                  sg.Input(last_src, key='-SOURCE-', enable_events=True),
-                  sg.FolderBrowse(initial_folder=last_src)]
+                  sg.Input(configs["gui_last_src"], key='-SOURCE-', enable_events=True),
+                  sg.FolderBrowse(initial_folder=configs["gui_last_src"])]
 
-    last_dest = sg.user_settings_get_entry('-last dest-', '')
     select_dest = [sg.Text("     Destination", size=(13, 1)),
-                   sg.Input(last_dest, key='-DEST-'),
-                   sg.FolderBrowse(initial_folder=last_dest)]
-
-    select_boxart = [sg.Text("     Boxart", size=(13, 1)),
-                     sg.Input(sg.user_settings_get_entry('-last boxarts-', ''),
-                     key='-BOXART-', size=(57, 1))]
-
-    select_img = [sg.Text("     Image", size=(13, 1)),
-                  sg.Input(sg.user_settings_get_entry('-last images-', ''),
-                  key='-IMAGE-', size=(57, 1))]
-
-    select_thumb = [sg.Text("     Thumbnail", size=(13, 1)),
-                    sg.Input(sg.user_settings_get_entry('-last thumbs-', ''),
-                    key='-THUMB-', size=(57, 1))]
-
-    select_marc = [sg.Text("     Marquee", size=(13, 1)),
-                   sg.Input(sg.user_settings_get_entry('-last marquees-', ''),
-                   key='-MARQUEE-', size=(57, 1))]
-
-    select_vid = [sg.Text("     Video", size=(13, 1)),
-                  sg.Input(sg.user_settings_get_entry('-last videos-', ''),
-                  key='-VIDEO-', size=(57, 1))]
-
-    subsystems = [sg.Text("     Subsystems", size=(13, 1)),
-                  sg.Multiline(sg.user_settings_get_entry('-last subsystems-', ''),
-                  key='-SUBSYSLIST-', size=(55, 6))]
-    layout = [
-        [sg.Text("Choose Operation.", font=("Helvetica", 12))],
-        operation,
-        [sg.Text("Set Source and Destination Paths.", font=("Helvetica", 12))],
-        select_src,
-        select_dest,
-        [sg.Text("Source Media Folder Names.", font=("Helvetica", 12)),
+                   sg.Input(configs["gui_last_dest"], key='-DEST-'),
+                   sg.FolderBrowse(initial_folder=configs["gui_last_dest"])]
+    
+    src_mdir_names = [sg.Text("Source Media Folder Names.", font=("Helvetica", 12)),
          sg.Text("(Comma separated, no spaces after comma)",
                  font=("Helvetica", 8),
-                 text_color='dark gray')],
+                 text_color='dark gray')]
+
+    select_boxart = [sg.Text("     Boxart", size=(13, 1)),
+                     sg.Input(",".join(configs["src_media_dirs_list"]["boxart"]), key='-BOXART-', size=(57, 1))]
+
+    select_img = [sg.Text("     Image", size=(13, 1)),
+                  sg.Input(",".join(configs["src_media_dirs_list"]["image"]), key='-IMAGE-', size=(57, 1))]
+
+    select_thumb = [sg.Text("     Thumbnail", size=(13, 1)),
+                    sg.Input(",".join(configs["src_media_dirs_list"]["thumbnail"]), key='-THUMB-', size=(57, 1))]
+
+    select_marc = [sg.Text("     Marquee", size=(13, 1)),
+                   sg.Input(",".join(configs["src_media_dirs_list"]["marquee"]), key='-MARQUEE-', size=(57, 1))]
+
+    select_vid = [sg.Text("     Video", size=(13, 1)),
+                  sg.Input(",".join(configs["src_media_dirs_list"]["video"]), key='-VIDEO-', size=(57, 1))]
+    
+    act_buttons = [sg.Button('Ok'), sg.Button('Exit')]
+    layout = [
+        choose_operation,
+        operation,
+        set_src_dest,
+        select_src,
+        select_dest,
+        src_mdir_names,
         select_boxart,
         select_img,
         select_thumb,
         select_marc,
         select_vid,
-        [sg.Text("SubSystems Names.", font=("Helvetica", 12)),
-         sg.Text("(Comma separated, no spaces after comma, multiple lines)",
-                 font=("Helvetica", 8),
-                 text_color='dark gray')],
-        subsystems,
-        [sg.Button('Ok'), sg.Button('Exit')],
+        act_buttons,
     ]
 
     return sg.Window('SBFRM V0.4.5', layout, enable_close_attempted_event=True,
-                     finalize=True,
-                     location=sg.user_settings_get_entry('-location-',
-                                                         (None, None)))
+                     finalize=True, icon="logo.ico",
+                     location=sg.user_settings_get_entry('-location-', (300, 300)))
 
 
 def make_progress_window(location):
@@ -100,28 +94,7 @@ def make_progress_window(location):
 
 
 def update_collection(collection, values, gui):
-    """Make Description"""
-    mdirs = {"boxart": values['-BOXART-'],
-             "image": values['-IMAGE-'],
-             "marquee": values['-THUMB-'],
-             "thumbnail": values['-MARQUEE-'],
-             "video": values['-VIDEO-']}
-    for key, value in mdirs.items():
-        while ', ' in value:
-            value = value.replace(', ', ',')
-        while value[-1] in [',', ' ']:
-            value = value[:-1]
-        mdirs[key] = value.split(',')
-
-    subsyslist = values['-SUBSYSLIST-'].splitlines()
-    for subsysid, value in enumerate(subsyslist):
-        while ', ' in value:
-            value = value.replace(', ', ',')
-        while value[-1] in [',', ' ']:
-            value = value[:-1]
-        subsyslist[subsysid] = value.split(',')
-
-    subsyslist = [s for ls in subsyslist for s in ls]
+    """Make Description"""    
     print(60*'='+'\n', 'Update Collection:')
 
     sys_paths = collection.list_systems(values['-SOURCE-'])
@@ -133,8 +106,7 @@ def update_collection(collection, values, gui):
         gui.write_event_value('-PROGRESS_SYSTEMS-', [sysid, progress_base])
         src_path = join(values['-SOURCE-'], sys_path)
         dest_path = join(values['-DEST-'], sys_path)
-        collection.update_sys_from(dest_path, src_path, mdirs,
-                                   subsyslist=subsyslist, gui=gui)
+        collection.update_sys_from(dest_path, src_path)
 
     gui.write_event_value('-UPDATE_COLLECTION_END-', '')
 
@@ -143,36 +115,14 @@ def update_sys_from(collection, values, gui):
     """Make Description"""
     print(60*'='+'\n', 'Update System:', values['-DEST-'])
 
-    mdirs = {"boxart": values['-BOXART-'],
-             "image": values['-IMAGE-'],
-             "marquee": values['-THUMB-'],
-             "thumbnail": values['-MARQUEE-'],
-             "video": values['-VIDEO-']}
-    for key, value in mdirs.items():
-        while ', ' in value:
-            value = value.replace(', ', ',')
-        while value[-1] in [',', ' ']:
-            value = value[:-1]
-        mdirs[key] = value.split(',')
-
-    subsyslist = values['-SUBSYSLIST-'].splitlines()
-    for subsysid, value in enumerate(subsyslist):
-        while ', ' in value:
-            value = value.replace(', ', ',')
-        while value[-1] in [',', ' ']:
-            value = value[:-1]
-        subsyslist[subsysid] = value.split(',')
-    subsyslist = [s for ls in subsyslist for s in ls]
-
-    collection.update_sys_from(values['-DEST-'], values['-SOURCE-'],
-                               mdirs, subsyslist=subsyslist, gui=gui)
+    collection.update_sys_from(values['-DEST-'], values['-SOURCE-'])
 
     gui.write_event_value('-UPDATE_SYSTEM_END-', '')
 
 
 def main():
     """Make Description."""
-    sg.theme('DarkAmber')  # please make your windows colorful
+    sg.theme('DarkAmber') 
 
     window1, window2 = make_window(), None
     col = None
@@ -189,16 +139,19 @@ def main():
 
             last = values['-UPDATE_SYSTEM-']
             sg.user_settings_set_entry('-last up_sys-', last)
-
-            sg.user_settings_set_entry('-last source-', values['-SOURCE-'])
-            sg.user_settings_set_entry('-last dest-', values['-DEST-'])
-            sg.user_settings_set_entry('-last boxarts-', values['-BOXART-'])
-            sg.user_settings_set_entry('-last images-', values['-IMAGE-'])
-            sg.user_settings_set_entry('-last thumbs-', values['-THUMB-'])
-            sg.user_settings_set_entry('-last marquees-', values['-MARQUEE-'])
-            sg.user_settings_set_entry('-last videos-', values['-VIDEO-'])
-            sg.user_settings_set_entry('-last subsystems-',
-                                       values['-SUBSYSLIST-'])
+            
+            configs["gui_last_src"] = values['-SOURCE-']
+            configs["gui_last_dest"] = values['-DEST-']
+            configs["src_media_dirs_list"]["boxart"] = values['-BOXART-'].split(",")
+            configs["src_media_dirs_list"]["image"] = values['-IMAGE-'].split(",")
+            configs["src_media_dirs_list"]["marquee"] = values['-MARQUEE-'].split(",")
+            configs["src_media_dirs_list"]["thumbnail"] = values['-THUMB-'].split(",")
+            configs["src_media_dirs_list"]["video"] = values['-VIDEO-'].split(",")
+            jsonString = json.dumps(configs, indent=4)
+            jsonFile = open("config.json", "w")
+            jsonFile.write(jsonString)
+            jsonFile.close()
+            
             window.close()
             if window == window2:
                 window2 = None
@@ -221,7 +174,7 @@ def main():
             xpos += 30
             ypos += 300
             window2 = make_progress_window((xpos, ypos))
-            col = Collection()
+            col = Collection(window)
             thread = threading.Thread(
                 target=update_collection, args=(col, values, window),
                 daemon=True)