+++ src/game.py
@@ -2,6 +2,8 @@
 
 import xml.etree.ElementTree as ET
 from os.path import isfile, basename, dirname, join, splitext
+
+from src.fileutils import configs, get_file_name
 from src.terminalutils import text_colored as tc
 
 
@@ -55,28 +57,23 @@ class Game:
             sout += "      " + tc('green', key) + ": " + str(value) + "\n"
         return sout
 
-    def load(self, paths, gamelist_path, media_dirs, subsys=None):
+    def load(self, paths, gamelist_path):
         """Make Description."""
-        self.set_paths(paths, media_dirs, subsys=subsys)
+        self.set_paths(paths)
         self.load_xml_paths(gamelist_path)
         self.load_xml_info(gamelist_path)
 
-    def set_paths(self, paths, media_dirs, subsys=None):
+    def set_paths(self, paths):
         """Make Description."""
         for key, value in paths.items():
             if value is not None and isfile(value):
                 if key == 'path':
                     path = basename(value)
-                    if subsys is not None:
-                        path = join(subsys, path)
                     self.paths[key] = "./" + path
                 else:
-                    if key in media_dirs.keys():
+                    if key in configs["dest_media_dirs_names"].keys():
                         path = basename(value)
                         dir_name = dirname(value)
-                        if subsys is not None:
-                            path = join(subsys, path)
-                            dir_name = dirname(dir_name)
                         dir_name = basename(dir_name)
                         if isfile(value):
                             self.paths[key] = "./" + join(dir_name, path)
@@ -112,8 +109,10 @@ class Game:
             for game in tree.getroot():
                 if game.find('path') is not None:
                     if game.find('path').text is not None:
-                        name_a = splitext(game.find('path').text)[0]
-                        name_b = splitext(self.paths['path'])[0]
+                        # name_a = splitext(game.find('path').text)[0]
+                        # name_b = splitext(self.paths['path'])[0]
+                        name_a = get_file_name(game.find('path').text)
+                        name_b = get_file_name(self.paths['path'])
                         if name_a == name_b:
                             games_xml.append(game)
         return games_xml
@@ -123,7 +122,12 @@ class Game:
         game_xml = ET.Element('game')
         for key, value in self.paths.items():
             subelement = ET.SubElement(game_xml, key)
+            if value is not None:
+                while "./" in value:
+                    value = value.replace("./", "")
+                value = "./"+value
             subelement.text = value
+            
         for key, value in self.info.items():
             subelement = ET.SubElement(game_xml, key)
             subelement.text = value